feat(environment): [#58038] added 'host.docker.internal' to ALLOWED_HOSTS
We need to change the network_mode to 'bride' on MacOS to do development,
since docker runs in a VM on macos, which also means we need to use port-forwarding.
Because of this we need to use 'host.docker.internal' instead of 'localhost'
to access DB, UI etc.
NOTE: Windows runs docker in a VM as well