fix: [#62435] Allow multiple users with same username
If an email is provided in SAML data, we use it for the `User.username`,
which must be unique (this assumes a 1:1 relationship between email
adresses and CPRs.)

If an email is not provided, we use the first name, last name, and CVR
(if provided) to construct a unique `User.username`.

If the constructed username is not unique on its own, we add a numeric
suffix starting from 1 and counting up. E.g. "Name Surname" is turned
into "Name Surname (43)" if "Name Surname (42)" already exists.

The unittest provided in this MR is stateful, and pollutes the "main"
database used by the application instance. It creates 3 new `User`
objects on each run. This is not assumed to be a blocking issue, as the
unittests are not intended to run in a production environment.